LPD-50711 Extract common code into a base class
