+++ modules/apps/site/site-cms-site-initializer-test/src/testIntegration/java/com/liferay/site/cms/site/initializer/internal/display/context/test/ContentsSectionDisplayContextTest.java
@@ -8,55 +8,37 @@ package com.liferay.site.cms.site.initializer.internal.display.context.test;
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.fragment.renderer.FragmentRenderer;
 import com.liferay.frontend.data.set.model.FDSActionDropdownItem;
-import com.liferay.frontend.taglib.clay.servlet.taglib.util.CreationMenu;
-import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem;
 import com.liferay.object.constants.ObjectDefinitionConstants;
 import com.liferay.object.constants.ObjectEntryFolderConstants;
-import com.liferay.object.field.util.ObjectFieldUtil;
 import com.liferay.object.model.ObjectDefinition;
 import com.liferay.object.model.ObjectFolder;
-import com.liferay.object.service.ObjectDefinitionLocalService;
-import com.liferay.object.service.ObjectFolderLocalService;
-import com.liferay.object.test.util.ObjectDefinitionTestUtil;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.petra.string.StringUtil;
-import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
-import com.liferay.portal.kernel.service.CompanyLocalService;
 import com.liferay.portal.kernel.test.ReflectionTestUtil;
 import com.liferay.portal.kernel.test.TestInfo;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
-import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
 import com.liferay.portal.kernel.test.rule.Sync;
-import com.liferay.portal.kernel.test.util.GroupTestUtil;
-import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.LinkedHashMapBuilder;
 import com.liferay.portal.kernel.util.LocaleUtil;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.test.rule.FeatureFlags;
 import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
 import javax.servlet.http.HttpServletRequest;
 
 import org.junit.Assert;
-import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
 
 /**
@@ -64,7 +46,8 @@ import org.springframework.mock.web.MockHttpServletResponse;
  */
 @RunWith(Arquillian.class)
 @Sync
-public class ContentsSectionDisplayContextTest {
+public class ContentsSectionDisplayContextTest
+	extends BaseSectionDisplayContextTestCase {
 
 	@ClassRule
 	@Rule
@@ -72,11 +55,6 @@ public class ContentsSectionDisplayContextTest {
 		new AggregateTestRule(
 			new LiferayIntegrationTestRule(),
 			PermissionCheckerMethodTestRule.INSTANCE);
-
-	@Before
-	public void setUp() throws Exception {
-		_group = GroupTestUtil.addGroup();
-	}
 
 	@Test
 	public void testGetAPIURL() throws Exception {
@@ -103,49 +81,57 @@ public class ContentsSectionDisplayContextTest {
 			"folder", StringPool.BLANK
 		).put(
 			"Basic Web Content",
-			_getHref(
-				_objectDefinitionLocalService.
+			getHref(
+				objectDefinitionLocalService.
 					fetchObjectDefinitionByExternalReferenceCode(
 						"L_BASIC_WEB_CONTENT", TestPropsValues.getCompanyId()))
 		).build();
 
-		_testGetCreationMenu(expectedResultMap);
+		testGetCreationMenu(
+			ReflectionTestUtil.invoke(
+				_getContentsSectionDisplayContext(getMockHttpServletRequest()),
+				"getCreationMenu", new Class<?>[0]),
+			expectedResultMap);
 
 		ObjectFolder objectFolder =
-			_objectFolderLocalService.fetchObjectFolderByExternalReferenceCode(
+			objectFolderLocalService.fetchObjectFolderByExternalReferenceCode(
 				"L_CMS_CONTENT_STRUCTURES", TestPropsValues.getCompanyId());
 
-		ObjectDefinition objectDefinition = _addCustomObjectDefinition(
+		ObjectDefinition objectDefinition = addCustomObjectDefinition(
 			objectFolder.getObjectFolderId(), true, true,
 			ObjectDefinitionConstants.SCOPE_SITE,
 			WorkflowConstants.STATUS_APPROVED);
 
 		expectedResultMap.put(
 			objectDefinition.getLabel(LocaleUtil.US),
-			_getHref(objectDefinition));
+			getHref(objectDefinition));
 
-		_addCustomObjectDefinition(
+		addCustomObjectDefinition(
 			ObjectEntryFolderConstants.PARENT_OBJECT_ENTRY_FOLDER_ID_DEFAULT,
 			false, true, ObjectDefinitionConstants.SCOPE_SITE,
 			WorkflowConstants.STATUS_APPROVED);
-		_addCustomObjectDefinition(
+		addCustomObjectDefinition(
 			objectFolder.getObjectFolderId(), false, true,
 			ObjectDefinitionConstants.SCOPE_SITE,
 			WorkflowConstants.STATUS_APPROVED);
-		_addCustomObjectDefinition(
+		addCustomObjectDefinition(
 			objectFolder.getObjectFolderId(), true, false,
 			ObjectDefinitionConstants.SCOPE_SITE,
 			WorkflowConstants.STATUS_APPROVED);
-		_addCustomObjectDefinition(
+		addCustomObjectDefinition(
 			objectFolder.getObjectFolderId(), true, true,
 			ObjectDefinitionConstants.SCOPE_COMPANY,
 			WorkflowConstants.STATUS_APPROVED);
-		_addCustomObjectDefinition(
+		addCustomObjectDefinition(
 			objectFolder.getObjectFolderId(), true, true,
 			ObjectDefinitionConstants.SCOPE_SITE,
 			WorkflowConstants.STATUS_DRAFT);
 
-		_testGetCreationMenu(expectedResultMap);
+		testGetCreationMenu(
+			ReflectionTestUtil.invoke(
+				_getContentsSectionDisplayContext(getMockHttpServletRequest()),
+				"getCreationMenu", new Class<?>[0]),
+			expectedResultMap);
 	}
 
 	@Test
@@ -177,64 +163,9 @@ public class ContentsSectionDisplayContextTest {
 		Assert.assertEquals("item", fdsActionDropdownItem.get("type"));
 	}
 
-	private ObjectDefinition _addCustomObjectDefinition(
-			long objectFolderId, boolean active, boolean enableObjectEntryDraft,
-			String scope, int status)
-		throws Exception {
-
-		ObjectDefinition objectDefinition =
-			_objectDefinitionLocalService.addCustomObjectDefinition(
-				TestPropsValues.getUserId(), objectFolderId, null, false, false,
-				true, true, enableObjectEntryDraft,
-				Collections.singletonMap(
-					LocaleUtil.getDefault(), RandomTestUtil.randomString()),
-				ObjectDefinitionTestUtil.getRandomName(), null, null,
-				Collections.singletonMap(
-					LocaleUtil.getDefault(), RandomTestUtil.randomString()),
-				true, scope, ObjectDefinitionConstants.STORAGE_TYPE_DEFAULT,
-				Collections.emptyList(),
-				Collections.singletonList(
-					ObjectFieldUtil.createObjectField(
-						"Text", "String", true, true, null,
-						RandomTestUtil.randomString(),
-						"x" + RandomTestUtil.randomString(), false)));
-
-		if (status == WorkflowConstants.STATUS_DRAFT) {
-			return objectDefinition;
-		}
-
-		objectDefinition =
-			_objectDefinitionLocalService.publishCustomObjectDefinition(
-				TestPropsValues.getUserId(),
-				objectDefinition.getObjectDefinitionId());
-
-		return _objectDefinitionLocalService.updateCustomObjectDefinition(
-			objectDefinition.getExternalReferenceCode(),
-			objectDefinition.getObjectDefinitionId(),
-			objectDefinition.getAccountEntryRestrictedObjectFieldId(),
-			objectDefinition.getDescriptionObjectFieldId(),
-			objectDefinition.getObjectFolderId(),
-			objectDefinition.getTitleObjectFieldId(),
-			objectDefinition.isAccountEntryRestricted(), active,
-			objectDefinition.getClassName(),
-			objectDefinition.isEnableCategorization(),
-			objectDefinition.isEnableComments(),
-			objectDefinition.isEnableFriendlyURLCustomization(),
-			objectDefinition.isEnableIndexSearch(),
-			objectDefinition.isEnableLocalization(),
-			objectDefinition.isEnableObjectEntryDraft(),
-			objectDefinition.isEnableObjectEntryHistory(),
-			objectDefinition.getLabelMap(), objectDefinition.getName(),
-			objectDefinition.getPanelAppOrder(),
-			objectDefinition.getPanelCategoryKey(),
-			objectDefinition.isPortlet(), objectDefinition.getPluralLabelMap(),
-			objectDefinition.getScope(), objectDefinition.getStatus(),
-			Collections.emptyList());
-	}
-
 	private String _getAPIURL() throws Exception {
 		return ReflectionTestUtil.invoke(
-			_getContentsSectionDisplayContext(_getMockHttpServletRequest()),
+			_getContentsSectionDisplayContext(getMockHttpServletRequest()),
 			"getAPIURL", new Class<?>[0]);
 	}
 
@@ -258,104 +189,21 @@ public class ContentsSectionDisplayContextTest {
 		throws Exception {
 
 		return ReflectionTestUtil.invoke(
-			_getContentsSectionDisplayContext(_getMockHttpServletRequest()),
+			_getContentsSectionDisplayContext(getMockHttpServletRequest()),
 			"getFDSActionDropdownItems", new Class<?>[0]);
-	}
-
-	private String _getHref(ObjectDefinition objectDefinition) {
-		StringBundler sb = new StringBundler(4);
-
-		sb.append("/cms/add_structured_content_item?groupId=");
-		sb.append(_group.getGroupId());
-		sb.append("&objectDefinitionId=");
-		sb.append(objectDefinition.getObjectDefinitionId());
-
-		return sb.toString();
-	}
-
-	private HttpServletRequest _getMockHttpServletRequest() throws Exception {
-		HttpServletRequest httpServletRequest = new MockHttpServletRequest();
-
-		httpServletRequest.setAttribute(
-			WebKeys.THEME_DISPLAY, _getThemeDisplay(httpServletRequest));
-
-		return httpServletRequest;
 	}
 
 	private String[] _getObjectDefinitionFolderExternalReferenceCodes()
 		throws Exception {
 
 		return ReflectionTestUtil.invoke(
-			_getContentsSectionDisplayContext(_getMockHttpServletRequest()),
+			_getContentsSectionDisplayContext(getMockHttpServletRequest()),
 			"getObjectDefinitionFolderExternalReferenceCodes", new Class<?>[0]);
 	}
-
-	private ThemeDisplay _getThemeDisplay(HttpServletRequest httpServletRequest)
-		throws Exception {
-
-		ThemeDisplay themeDisplay = new ThemeDisplay();
-
-		themeDisplay.setCompany(
-			_companyLocalService.getCompany(TestPropsValues.getCompanyId()));
-		themeDisplay.setPermissionChecker(
-			PermissionThreadLocal.getPermissionChecker());
-		themeDisplay.setRealUser(TestPropsValues.getUser());
-		themeDisplay.setRequest(httpServletRequest);
-		themeDisplay.setScopeGroupId(_group.getGroupId());
-		themeDisplay.setSiteGroupId(_group.getGroupId());
-		themeDisplay.setURLCurrent("http://localhost:8080/currentURL");
-		themeDisplay.setUser(TestPropsValues.getUser());
-
-		return themeDisplay;
-	}
-
-	private void _testGetCreationMenu(Map<String, String> expectedResultMap)
-		throws Exception {
-
-		CreationMenu creationMenu = ReflectionTestUtil.invoke(
-			_getContentsSectionDisplayContext(_getMockHttpServletRequest()),
-			"getCreationMenu", new Class<?>[0]);
-
-		List<DropdownItem> dropdownItems = (List<DropdownItem>)creationMenu.get(
-			"primaryItems");
-
-		Assert.assertEquals(
-			dropdownItems.toString(), expectedResultMap.size(),
-			dropdownItems.size());
-
-		int index = 0;
-
-		for (Map.Entry<String, String> entry : expectedResultMap.entrySet()) {
-			DropdownItem dropdownItem = dropdownItems.get(index);
-
-			Assert.assertEquals(entry.getKey(), dropdownItem.get("label"));
-
-			if (Validator.isNull(entry.getValue())) {
-				Assert.assertNull(dropdownItem.get("href"));
-			}
-			else {
-				Assert.assertEquals(entry.getValue(), dropdownItem.get("href"));
-			}
-
-			index++;
-		}
-	}
-
-	@Inject
-	private CompanyLocalService _companyLocalService;
 
 	@Inject(
 		filter = "component.name=com.liferay.site.cms.site.initializer.internal.fragment.renderer.ContentsSectionFragmentRenderer"
 	)
 	private FragmentRenderer _fragmentRenderer;
-
-	@DeleteAfterTestRun
-	private Group _group;
-
-	@Inject
-	private ObjectDefinitionLocalService _objectDefinitionLocalService;
-
-	@Inject
-	private ObjectFolderLocalService _objectFolderLocalService;
 
 }
